backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Caching
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.2.0

# Logging & Monitoring
structlog==24.1.0

# CORS
python-cors==1.0.0

# PDF & Excel Generation
reportlab==4.0.7
openpyxl==3.1.2
Pillow==10.1.0
